[Feature][FC-131] Rename folder
