+++ back/src/utils/constants/preferences/preferences.constants.ts
@@ -1,0 +1,15 @@
+export const defaultPreference = {
+    id: "3a9975f8-f34c-4a07-bbff-ab8a9b2e6309",
+    theme: "DEFAULT",
+    language: "FRENCH"
+}
+
+export const defaultPreferences = [{
+    id: "3a9975f8-f34c-4a07-bbff-ab8a9b2e6309",
+    theme: "DEFAULT",
+    language: "FRENCH"
+}, {
+    id: "2cd08d05-c7b3-4550-aa02-aeddf963411c",
+    theme: "DARK",
+    language: "FRENCH"
+}]
