+++ back/src/utils/seeders/preferences/preferences.seeder.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { Seeder } from "nestjs-seeder";
 import { Preference } from "src/schemas/preference.schema";
-import { preferences } from "./preferences.fixtures";
+import { defaultPreference, defaultPreferences } from "src/utils/constants/preferences/preferences.constants";
 
 @Injectable()
 export class PreferencesSeeder implements Seeder {
@@ -9,7 +9,7 @@ export class PreferencesSeeder implements Seeder {
 
   async seed(): Promise<any> {
     try {
-      return await Preference.bulkCreate(preferences);
+      return await Preference.bulkCreate(defaultPreferences);
     } catch (e) {
       console.log(e)
     }
