+++ back/src/utils/types/preferences/preferences.types.ts
@@ -1,0 +1,7 @@
+export type LANGUAGES = "FRENCH" | "ENGLISH"
+export type THEMES = "DEFAULT" | "DARK"
+
+export type PreferenceType = {
+    theme: THEMES,
+    language: LANGUAGES
+}
